+++ CHANGELOG.md
@@ -7,9 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 - Added
-  - Listing webhooks
+  - Listing, creating, and deleting webhooks
   - Viewing details of a webhook
-  - Creating webhooks
 
 ## [1.1.0] - 2019-10-07